# Conflicts:
#	API/pom.xml
#	IDE/src/main/java/org/sikuli/ide/SikulixIDE.java

+++ API/src/main/java/org/sikuli/basics/Settings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2018, sikuli.org, sikulix.com - MIT license
+ * Copyright (c) 2010-2019, sikuli.org, sikulix.com - MIT license
  */
 package org.sikuli.basics;
 
@@ -138,7 +138,7 @@ public class Settings {
   public static boolean Highlight = false;
   public static float DefaultHighlightTime = 2f;
   public static String DefaultHighlightColor = "RED";
-  public static float WaitAfterHighlight = 0.3f;
+  public static double WaitAfterHighlight = 0.3;
 
   public static boolean ActionLogs = true;
   public static boolean InfoLogs = true;
